+++ packages/selectors/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "sideEffects": false,
   "type": "module",
-  "description": "Shared selector matching, argument, and replay semantics for agent-device. `.` is string-only; `./ast` is the published parser surface behind `agent-device/selectors`; `./engine` is the resolve/list surface reserved for the selector-pipeline owner (R19); `./parameterized-recorded-fill` parameterizes recorded fill payloads against their selectors; the interaction-resolution subpaths (`./selector-pipeline`, `./interaction-targeting`, `./interaction-touch-point`, `./press-retarget`, `./absence-observation*`, …) host the engine execution surface owned by the pipeline.",
+  "description": "Shared selector matching, argument, and replay semantics for agent-device. `.` is string-only; `./ast` is the published parser surface behind `agent-device/selectors`; `./engine` is the resolve/list surface reserved for the selector-pipeline owner (R19); `./parameterized-recorded-fill` parameterizes recorded fill payloads against their selectors; the interaction-resolution subpaths (`./selector-pipeline`, `./interaction-targeting`, `./interaction-touch-point`, `./press-retarget`, `./absence-observation*`, …) host the engine execution surface owned by the pipeline; `./test-fixtures` is the canonical home for snapshot/geometry test fixtures this package and root tests both build on.",
   "dependencies": {
     "@agent-device/ad-script": "workspace:*",
     "@agent-device/capture-kit": "workspace:*",
@@ -71,6 +71,10 @@
     "./selector-pipeline-policy": {
       "types": "./src/selector-pipeline-policy.ts",
       "default": "./src/selector-pipeline-policy.ts"
+    },
+    "./test-fixtures": {
+      "types": "./src/test-fixtures.ts",
+      "default": "./src/test-fixtures.ts"
     },
     "./touch-semantics": {
       "types": "./src/touch-semantics.ts",
