--- packages/selectors/src/internal/__tests__/snapshot-builders.ts
@@ -1,19 +1,0 @@
-import {
-  attachRefs,
-  type RawSnapshotNode,
-  type SnapshotState,
-  type SnapshotStateProvenance,
-} from '@agent-device/kernel/snapshot';
-
-export function makeSnapshotState(
-  raw: RawSnapshotNode[],
-  // The provenance pair stays correlated: overrides carry it as one value, never as two
-  // independently typed fields.
-  overrides?: Omit<Partial<SnapshotState>, 'backend' | 'producer'> & SnapshotStateProvenance,
-): SnapshotState {
-  return {
-    nodes: attachRefs(raw),
-    createdAt: Date.now(),
-    ...overrides,
-  };
-}
