+++ src/commands/replay/ad-script-round-trip.test.ts
@@ -1,7 +1,8 @@
 import fc from 'fast-check';
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
-import { PROPERTY_RUNS, replayScriptArb } from '../../__tests__/test-utils/property-arbitraries.ts';
+import { replayScriptArb } from '../../__tests__/test-utils/property-arbitraries.ts';
+import { PROPERTY_RUNS } from '@agent-device/selectors/test-fixtures';
 import {
   formatPortableActionLine,
   formatTargetAnnotationLines,
