+++ src/daemon/__tests__/post-gesture-stabilization-fixtures.ts
@@ -1,5 +1,5 @@
 import { ANDROID_EMULATOR, IOS_SIMULATOR } from '../../__tests__/test-utils/device-fixtures.ts';
-import { makeSnapshotState } from '../../__tests__/test-utils/snapshot-builders.ts';
+import { makeSnapshotState } from '@agent-device/selectors/test-fixtures';
 import type { SessionState } from '../session-state.ts';
 
 /**
