+++ src/daemon/__tests__/recording-gestures.test.ts
@@ -5,7 +5,7 @@ import {
   recordTouchVisualizationEvent,
 } from '../recording-gestures.ts';
 import { makeIosSession, makeMacOsSession } from '../../__tests__/test-utils/session-factories.ts';
-import { makeSnapshotState } from '../../__tests__/test-utils/snapshot-builders.ts';
+import { makeSnapshotState } from '@agent-device/selectors/test-fixtures';
 import { makeTestScreenRecordingResource } from '../../__tests__/test-utils/screen-recording-live-handle.ts';
 import type { ScreenRecordingLiveSnapshot } from '@agent-device/contracts/screen-recording-runtime';
 import { TVOS_SIMULATOR } from '../../__tests__/test-utils/device-fixtures.ts';
