+++ src/daemon/interaction/internal/__tests__/find-match-ranking.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import { test } from 'vitest';
 import type { RawSnapshotNode, SnapshotState } from '@agent-device/kernel/snapshot';
-import { makeSnapshotState } from '../../../../__tests__/test-utils/snapshot-builders.ts';
+import { makeSnapshotState } from '@agent-device/selectors/test-fixtures';
 import { preferOnscreenMatches } from '../find-match-ranking.ts';
 
 const VIEWPORT = { x: 0, y: 0, width: 390, height: 844 };
